+++ docs/CHANGELOG.rst
@@ -10,6 +10,11 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 Unreleased
 ==========
 
+Added
+-----
+- Add ``run_process`` method to ``Process`` to support triggering
+  of a new process from the running Python process
+
 Changed
 -------
 - Purge processes only not jet purged Data objects